+++ src/main/java/samryong/domain/item/contorller/ItemController.java
@@ -17,9 +17,9 @@ import samryong.domain.item.dto.ItemDTO.RecentItemResponseListDTO;
 import samryong.domain.item.dto.ItemDTO.WishListResponseDTO;
 import samryong.domain.item.service.category.CategoryService;
 import samryong.domain.item.service.item.ItemService;
+import samryong.domain.item.service.recentItem.RecentItemService;
 import samryong.domain.location.dto.LocationDTO.LocationRequestDTO;
 import samryong.domain.member.entity.Member;
-import samryong.domain.redis.service.RecentItemService;
 import samryong.global.annotation.AuthMember;
 import samryong.global.response.ApiResponse;
