+++ src/main/java/samryong/domain/item/service/item/ItemServiceImpl.java
@@ -27,11 +27,11 @@ import samryong.domain.item.repository.ItemRepository;
 import samryong.domain.item.repository.elastic.ItemElasticRepository;
 import samryong.domain.item.service.category.CategoryService;
 import samryong.domain.item.service.itemCategory.ItemCategoryService;
+import samryong.domain.item.service.recentItem.RecentItemService;
 import samryong.domain.location.dto.LocationDTO.LocationRequestDTO;
 import samryong.domain.location.entity.Location;
 import samryong.domain.location.service.LocationService;
 import samryong.domain.member.entity.Member;
-import samryong.domain.redis.service.RecentItemService;
 import samryong.domain.uuid.Uuid;
 import samryong.domain.uuid.UuidRepository;
 import samryong.global.aws.AmazonS3Manager;
